+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - `EnableRecordSupportAttribute` - assembly level attribute for injecting support for records
 - `EnableRequiredSupportAttribute` - assembly level attribute for injecting support for required members
 - `ArrayBuilder<T>.ToImmutableArray()` for converting the builder contents to an immutable array
+- `VariableNames.AppendGlobalPrefix` for appending `global::` to a span or string
 
 ## 1.1.0 - 2026-08-01